requirements.txt
# CreditIQ — Python Dependencies

# Core API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1

# LangChain & AI Agent
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
langchain-pinecone==0.1.1
openai==1.30.1

# Vector Database
pinecone-client==3.2.2

# PDF & Document Processing
pymupdf==1.24.3
pdfplumber==0.11.0
openpyxl==3.1.2
pandas==2.2.2
numpy==1.26.4

# Web Search / Crawling
tavily-python==0.3.3

# ML & Credit Scoring
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.45.1
imbalanced-learn==0.12.2
joblib==1.4.2

# GAN Synthetic Data
ctgan==0.10.1
sdv==1.17.3

# Time Series
prophet==1.1.5

# Graph Analytics
networkx==3.3

# Embeddings
sentence-transformers==2.7.0

# ML Additional
lightgbm==4.3.0

# Utilities
httpx==0.27.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
